pyproject.toml
[tool.poetry]
name = "pub-msl"
version = "0.1.0"
description = ""
authors = ["Stefan <stefan.heid@lmu.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
scikit-learn = "~1.5"
matplotlib = "3.10.0"
pandas = "^2.2"
numpy = "^1.26"
scipy = "~1.14"
jupyter = "^1.0.0"
seaborn = "^0.13"
tqdm = "^4.66"
#scikit-psl = { git = "http://github.com/trr318/scikit-psl", branch = "master" }
scikit-psl = { path = "../scikit-psl/", develop = true }
miss = { path = "../MISS/", develop = true }
joblib = "^1.4"
networkx = "~3.2"
optuna = "^4.0.0"
xgboost = "^2.1.1"
uncertainty-calibration = "0.1.4"
mapie = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
